+++ game_patch/rf/network.h
@@ -104,7 +104,7 @@ namespace rf
     static auto& CtfGetBlueFlagPlayer = AddrAsRef<Player*()>(0x00474E70);
     static auto& KickPlayer = AddrAsRef<void(Player *player)>(0x0047BF00);
     static auto& BanIp = AddrAsRef<void(const NwAddr& addr)>(0x0046D0F0);
-    static auto& MultiSetNextWeapon = AddrAsRef<void(int weapon_cls_id)>(0x0047FCA0);
+    static auto& MultiSetRequestedWeapon = AddrAsRef<void(int weapon_cls_id)>(0x0047FCA0);
     static auto& MultiChangeLevel = AddrAsRef<void(const char* filename)>(0x0047BF50);
     static auto& PingPlayer = AddrAsRef<void(rf::Player*)>(0x00484D00);