+++ CHANGELOG.md
@@ -55,6 +55,7 @@ Version 1.4.2
 - fix crash caused by explosion near dying player-controlled entity (RF bug)
 - fix unwanted fire after weapon switch from weapon with impact delay (RF bug)
 - fix wall-peeking (RF bug)
+- fix weapon being auto-switched to previous one after respawn even when auto-switch is disabled (RF bug)
 
 Version 1.4.1 (released 2019-12-07)
 -----------------------------------